+++ public/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <link rel="icon" href="%PUBLIC_URL%/fireboyreal.png"style="width: 120px; height: auto;" />
-    <script src="https://developers.kakao.com/sdk/js/kakao.js"></script>
+    <!-- <script src="https://developers.kakao.com/sdk/js/kakao.js"></script> -->
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <script
     type="text/javascript"
@@ -14,7 +14,7 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
+    <!-- <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" /> -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
     <title>Emotion Hell</title>
   </head>
